refactor(utils): move event forwarding functions to global utils
